requirement.txt
absl-py==0.11.0
asn1crypto==0.24.0
astunparse==1.6.3
Authlib==0.11
bcrypt==3.2.0
beautifulsoup4==4.8.0
cached-property==1.5.2
cachetools==3.1.1
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
Django==2.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Ext==0.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-SocketIO==4.2.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.18.2
gast==0.3.3
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
httplib2==0.13.1
idna==2.8
image==1.5.33
image-slicer==2.1.1
imageio==2.9.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mysqlclient==1.4.4
networkx==2.5
numpy==1.17.0
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.0.1
opencv-python==4.4.0.46
opt-einsum==3.3.0
panda==0.3.1
pandas==1.1.4
Pillow==7.2.0
pipenv==2018.11.26
protobuf==3.14.0
py-dateutil==2.2
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.9.0
python-socketio==4.3.0
pytz==2017.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
rfc3339==6.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
torch==1.5.1+cu101
torchaudio==0.7.0
torchvision==0.6.1+cu101
typing-extensions==3.7.4.3
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==15.1.0
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wrapt==1.12.1
WTForms==2.2.1
zipp==3.4.0
